cleaning up code and front-end improving
