+++ public/src/app.js
@@ -1,0 +1,49 @@
+// Create a Node.js application that is the beginning of a user management system. Your users are all saved in a "users.json" file, and you can currently do the following:
+// - search for users
+// - add new users to your users file.
+// - get your starter file here: users.json
+
+// Part 0
+// Create one route:
+// - route 1: renders a page that displays all your users.
+
+// Part 1
+// Create two more routes:
+// - route 2: renders a page that displays a form which is your search bar.
+// - route 3: takes in the post request from your form, then displays matching users on a new page. 
+// Users should be matched based on whether either their first or last name contains the input string.
+
+// Part 2
+// Create two more routes:
+// - route 4: renders a page with three forms on it (first name, last name, and email) that allows you 
+// to add new users to the users.json file.
+// - route 5: takes in the post request from the 'create user' form, then adds the user to the users.json file. 
+// Once that is complete, redirects to the route that displays all your users (from part 0).
+
+const server = require('./server');
+
+const reader = require('./json-reader');
+const writer = require('./json-writer');
+
+const jsonUser = './users.json';
+
+reader.readFiles(jsonUser,  (contents) => {
+	let newUser = '';
+	let input = process.argv[2];
+
+	for (var i = 0; i < contents.length; i++) {
+		let name = contents[i].firstname;
+		if (input === name && input !== undefined) {
+			return console.log(`yeeah ${name}, you're in the server!`);
+		} 
+	}
+
+	for (var i = 0; i < contents.length; i++) {
+		let name = contents[i].firstname;
+		if (input !== name && input !== undefined) {
+			return writer.writeFiles(contents, input, jsonUser);
+		} 
+	}
+});
+
+server.serverConf();
