+++ public/src/json-reader.js
@@ -1,0 +1,10 @@
+module.exports = {
+	readFiles: (filename, callbackFunction) => {
+		const fs = require("fs");
+		fs.readFile(filename,'utf-8' ,(err, data) => {	
+		 	if(err) throw err;
+		 	const contents = JSON.parse(data);
+		 	callbackFunction(contents);
+		});
+	}
+};
