+++ users.json
@@ -1,1 +1,1 @@
-[{"firstname":"Alexandre","lastname":"Marjan","email":"a.marjan@gmail.com"},{"firstname":"Ulysses","lastname":"Silvio","email":"silverspoon87@gmail.com"},{"firstname":"Viktor","lastname":"Gul","email":"viktor.rose@gmail.com"},{"firstname":"Kamalani","lastname":"Georgy","email":"georgy.porgy@gmail.com"},{"firstname":"Savino","lastname":"Rasola","email":"savinorasola90@gmail.com"},{"firstname":"Lloyd","lastname":"C","email":"lloyd@test.com"},{"firstname":"test","lastname":"testa","email":"test@testa.gmail.com"}]
+[{"firstname":"Alexandre","lastname":"Marjan","email":"a.marjan@gmail.com"},{"firstname":"Ulysses","lastname":"Silvio","email":"silverspoon87@gmail.com"},{"firstname":"Viktor","lastname":"Gul","email":"viktor.rose@gmail.com"},{"firstname":"Kamalani","lastname":"Georgy","email":"georgy.porgy@gmail.com"},{"firstname":"Savino","lastname":"Rasola","email":"savinorasola90@gmail.com"},{"firstname":"Lloyd","lastname":"C","email":"lloyd@test.com"},{"firstname":"test","lastname":"testa","email":"test@testa.gmail.com"},{"firstname":"joe","lastname":"boh","email":""},{"firstname":"test","lastname":"test4","email":""}]
