+++ views/pages/search.ejs
@@ -13,35 +13,14 @@
   <div class="jumbotron">
     <h1>Search User</h1>
     <p>Welcome to the managment system app</p>
-  </div>
-  <div class="jumbotron">
-  	<form action="" method="post" >
-		  <h4>Search:</h4>
-		  <input id="input_search" type="text" name="firstname" placeholder="Mickey"><br><br>
-        <ul id="suggestions" class="list-group"></ul>
-      <input type="submit" value="Submit">
-      
-      
-      <script>
-        $(document).ready(() => {
-          $("#input_search").keyup((e) => {          
-            
-            let name = $('#input_search').val();
-            if (name != '') {
-              // sendind data using post
-              let posting = $.post('/users', {name:name});
-              posting.done((data) => {
-                // console.log(data);
-                for(const el in data){
-                  console.log(el);
-                  $('#suggestions').append(`<li class="list-group-item" style="list-style:none;">${data[el]}</li>`);
-                }
-              })
-            }
-            $('#suggestions').empty();       
-          });
-        });
-      </script>
+    <form action="" method="post" >
+      <div class="container-1">
+  		  <input id="input_search" type="text" name="firstname" placeholder="Search...">
+        <button id="btn" type="submit" value="Submit">
+          <span class="icon"><i class="glyphicon glyphicon-search"></i></span>
+        </button>      
+        <ul id="suggestions" class="list-group" clear:both;"></ul>
+      </div>
 	  </form>
   </div>
 </main>
@@ -49,6 +28,26 @@
 <footer>
   <% include ../partials/footer %>
 </footer>
+
+<script>
+  $(document).ready(() => {
+    $("#input_search").keyup((e) => {          
+      let name = $('#input_search').val();
+      if (name != '') {
+        // sendind data using post
+        let posting = $.post('/users', {name:name});
+        posting.done((data) => {
+          // console.log(data);
+          for(const el in data){
+            console.log(el);
+            $('#suggestions').append(`<li class="list-group-item" style="list-style:none;">${data[el]}</li>`);
+          }
+        })
+      }
+      $('#suggestions').empty();       
+    });
+  });
+</script>
 
 </body>
 </html>
