+++ views/pages/usrAdd.ejs
@@ -16,12 +16,18 @@
   </div>
   <div class="jumbotron">
   	<form action="" method="post" >
-		  First name:<br>
-		  <input type="text" name="firstname" value=""><br>
-		  Last name:<br>
-		  <input type="text" name="lastname" value=""><br>
-      Email:<br>
-      <input type="email" name="Email" value=""><br><br>
+        <div class="add addName">
+          <p>First name:</p>
+          <input id="input_search" type="text" name="firstname" value="">
+        </div>
+        <div class="add addlname">
+          <p>Last name:</p>
+          <input id="input_search" type="text" name="lastname" value="">
+        </div>
+        <div class="add addEmail">
+          <p>Email:</p>
+          <input id="input_search" type="email" name="Email" value="">
+        </div>
 		  <input type="submit" value="Submit"><br>
 	  </form>
   </div>
