+++ views/partials/footer.ejs
@@ -1,2 +1,6 @@
 <!-- views/partials/footer.ejs -->
-<p class="text-center text-muted">© Copyright 2018 Savino Rasola</p>
+<div class="jumbotron">
+	<p class="text-center">© Copyright 2018 Savino Rasola</p>
+</div>
+
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
