+++ views/partials/head.ejs
@@ -1,10 +1,13 @@
 <!-- views/partials/head.ejs -->
+<!doctype html>
 <meta charset="UTF-8">
 <title>Managment system app</title>
 
 <!-- CSS (load bootstrap from a CDN) -->
 <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.2.0/css/bootstrap.min.css">
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+
+<link href="style/style.css" rel="stylesheet" type="text/css">
+
 <style>
     body    { padding-top:50px; }
 </style>
